transcriber/requirements.txt
flask
openai-whisper
srt
ffmpeg-python
